// frontend/src/api/models/add-species-request-body.ts
/* tslint:disable */
/* eslint-disable */
/**
 * OpenAPI definition
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: v0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



/**
 * 
 * @export
 * @interface AddSpeciesRequestBody
 */
export interface AddSpeciesRequestBody {
    /**
     * 
     * @type {string}
     * @memberof AddSpeciesRequestBody
     */
    'name': string;
    /**
     * 
     * @type {string}
     * @memberof AddSpeciesRequestBody
     */
    'fertilizationDose': AddSpeciesRequestBodyFertilizationDoseEnum;
    /**
     * 
     * @type {number}
     * @memberof AddSpeciesRequestBody
     */
    'fertilizationRoutine': number;
    /**
     * 
     * @type {number}
     * @memberof AddSpeciesRequestBody
     */
    'maxAge': number;
    /**
     * 
     * @type {string}
     * @memberof AddSpeciesRequestBody
     */
    'neededInsolation': AddSpeciesRequestBodyNeededInsolationEnum;
    /**
     * 
     * @type {number}
     * @memberof AddSpeciesRequestBody
     */
    'waterDose': number;
    /**
     * 
     * @type {number}
     * @memberof AddSpeciesRequestBody
     */
    'waterRoutine': number;
}

export const AddSpeciesRequestBodyFertilizationDoseEnum = {
    Low: 'LOW',
    Medium: 'MEDIUM',
    High: 'HIGH'
} as const;

export type AddSpeciesRequestBodyFertilizationDoseEnum = typeof AddSpeciesRequestBodyFertilizationDoseEnum[keyof typeof AddSpeciesRequestBodyFertilizationDoseEnum];
export const AddSpeciesRequestBodyNeededInsolationEnum = {
    Low: 'LOW',
    Medium: 'MEDIUM',
    High: 'HIGH'
} as const;

export type AddSpeciesRequestBodyNeededInsolationEnum = typeof AddSpeciesRequestBodyNeededInsolationEnum[keyof typeof AddSpeciesRequestBodyNeededInsolationEnum];